requirements.txt
kagglehub
numpy
opencv-python
geopandas
matplotlib
networkx
shapely
torch
torchmetrics
tqdm
hydra-core